moving metadata conversion
