+++ backend/tests/Designer.Tests/Factories/ModelFactory/JsonSchema2Metadata2CSharpTests.cs
@@ -11,6 +11,7 @@ using System.Xml.Schema;
 using System.Xml.Serialization;
 using Altinn.Studio.DataModeling.Converter.Csharp;
 using Altinn.Studio.DataModeling.Converter.Json.Strategy;
+using Altinn.Studio.DataModeling.Converter.Metadata;
 using Altinn.Studio.DataModeling.Converter.Xml;
 using Altinn.Studio.DataModeling.Json.Keywords;
 using Altinn.Studio.DataModeling.Metamodel;
