+++ backend/tests/Designer.Tests/Factories/ModelFactory/JsonSchemaToMetamodelConverterTests.cs
@@ -7,6 +7,7 @@ using System.Text.Unicode;
 using System.Xml.Schema;
 using Altinn.Studio.DataModeling.Converter.Csharp;
 using Altinn.Studio.DataModeling.Converter.Json.Strategy;
+using Altinn.Studio.DataModeling.Converter.Metadata;
 using Altinn.Studio.DataModeling.Converter.Xml;
 using Altinn.Studio.DataModeling.Json.Formats;
 using Altinn.Studio.DataModeling.Json.Keywords;
